Add an error summary component
This is intended to be placed over inputs in a form, and shows a
summary of all errors in the form.

It contains links to the input, so if you click it focus will
automatically jump there.

You can also set autoFocus={true} to have focus jump to the error summary
when you add it to the dom (recommended)