Support use of Breadcrumbs without reset.css
Prior to the change, when using Breadcrumbs
without reset.css, the chevron icon between
the links was not displayed.

The reset stylesheet contains a `content: ""`
rule which is applied to the `<li>` elements,
and this leads to the icon being shown. We've
just added the same rule to breadcrumbs.css,
and altered the display rule to align the links and
icons in the list items.

The rule from reset.css (sanitize.css):
https://github.com/csstools/sanitize.css/blob/61c85b6961f684d0124093d97475073bb08f557e/sanitize.css#L102-L105